Working on cleaning up type definitions in the web project
