+++ web/src/components/ui/sidebar-context.tsx
@@ -1,14 +1,5 @@
 import * as React from "react"
-
-export type SidebarContextProps = {
-  state: "expanded" | "collapsed"
-  open: boolean
-  setOpen: (open: boolean) => void
-  openMobile: boolean
-  setOpenMobile: (open: boolean) => void
-  isMobile: boolean
-  toggleSidebar: () => void
-}
+import type { SidebarContextProps } from "@/types"
 
 export const SidebarContext =
   React.createContext<SidebarContextProps | null>(null)
