+++ web/src/components/ui/sidebar.tsx
@@ -24,9 +24,9 @@ import {
 
 import {
   SidebarContext,
-  type SidebarContextProps,
   useSidebar,
 } from "@/components/ui/sidebar-context"
+import type { SidebarContextProps } from "@/types"
 
 const SIDEBAR_COOKIE_NAME = "sidebar_state"
 const SIDEBAR_COOKIE_MAX_AGE = 60 * 60 * 24 * 7
