+++ web/src/routes/_auth/movies/$id/play.tsx
@@ -39,7 +39,7 @@ import {
 import { showActionFailed } from "@/lib/toast-helpers";
 import { cn } from "@/lib/utils";
 import type { PlaySearchParams } from "@/types";
-import { playSearchSchema } from "@/types/movie-play";
+import { playSearchSchema } from "@/types/route-search";
 import { useAudioPlayerActions } from "@/hooks/useAudioPlayerActions";
 import { useVideoMediaSession } from "@/hooks/useVideoMediaSession";
 import { useVideoFullscreen } from "@/hooks/useVideoFullscreen";
