+++ web/src/routes/_auth/trailer.tsx
@@ -1,13 +1,6 @@
 import { createFileRoute } from "@tanstack/react-router";
-import { z } from "zod/mini";
 import { movieDetailsQueryOpts } from "@/lib/query-opts";
-
-const trailerSearchSchema = z.object({
-  mediaType: z.optional(z.enum(["movie", "tv"])),
-  mediaId: z.optional(z.coerce.number().check(z.int(), z.positive())),
-  videoKey: z.optional(z.string()),
-  returnTo: z.optional(z.string()),
-});
+import { trailerSearchSchema } from "@/types/route-search";
 
 export const Route = createFileRoute("/_auth/trailer")({
   validateSearch: trailerSearchSchema,
